fix: added forgot password base, fix super admin banner
