+++ frontend/src/modules/auth/pages/ForgotPasswordPage/components/BaseForgotPasswordPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './BaseForgotPasswordPage';
